+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mixins/mp_th2_file_edit_state_move_canvas_mixin.dart
@@ -77,7 +77,8 @@ mixin MPTH2FileEditStateMoveCanvasMixin on MPTH2FileEditState {
             !isCtrlPressed &&
             !isMetaPressed &&
             !isShiftPressed) {
-          th2FileEditController.performHideSelectedOrClearHidden();
+          th2FileEditController.hideElementController
+              .performHideSelectedOrClearHidden();
           keyProcessed = true;
         }
       case LogicalKeyboardKey.keyI: